[feat]#281: 코드 정리해요
